backend/requirements.txt
# Core dependencies (works on Python 3.11, 3.12, and 3.13)
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
pyyaml>=6.0.0
sqlparse>=0.5.0
sse-starlette>=1.6.0
faker>=22.0.0
networkx>=3.0
statsmodels>=0.14.0
pmdarima>=2.0.0
PyPDF2>=3.0.0
reportlab>=4.0.0
imbalanced-learn>=0.12.0
